Completed most of the movement related stories.
This required implementing a Game layer including an input layer and output
console - the CommandGenerator and GameDisplay respectively.  I also did quite
a bit of refactoring and introduced enumerations for the string literals that
were used in input+output and reworked error management so that it uses
exceptions instead of returning null or ignoring conditions.  This was
required to meet a number of the stories.

The following stories from Iteration One are now complete:
BatTransportRandomly, FallInPit, HearBats, HearPit,
NavigatePlayer.TestCanMoveThroughMap, NavigatePlayer.TestCanRest,
NavigatePlayer.TestCantMoveOutOfMap, NavigatePlayer.TestCommandAbbreviations,
NavigatePlayer.TestInvalidCommand.